+++ src/constants.ts
@@ -17,8 +17,20 @@ export const NETMASK_OCTETS = [
  * @hidden
  */
 export const IPV4_LENGTH = 32;
+/**
+ * @hidden
+ */
+export const IPV6_LENGTH = 128;
 
 /**
  * @hidden
  */
 export const IPV4_BYTES = 4;
+/**
+ * @hidden
+ */
+export const IPV6_BYTES = 16;
+/**
+ * @hidden
+ */
+export const IPV6_SEGMENT_LENGTH = 4;