+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Add `getDecimal` function to `Address4`.
+- Add `Address6` class, initial IPv6 support.
 ### Changed
 - Small optimization in broadcast address calculation.